Add jwt claims and temporal blocklist and user logout endpoint

+++ resources/item.py
@@ -1,6 +1,6 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from schemas import ItemSchema, ItemUpdateSchema, ResponseSchema
 from models import ItemModel
 from db import db
@@ -23,6 +23,9 @@ class Item(MethodView):
             abort(404, message="Item not found.")
 
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin Privileges Required.")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
